requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.38.15
botocore==1.38.15
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
google-auth==2.39.0
greenlet==3.2.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
passlib==1.7.4
pyasn1==0.4.8
pyasn1_modules==0.4.1
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
s3transfer==0.12.0
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
stripe==12.4.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
gunicorn==23.0.0

# PDF Processing Libraries
PyPDF2==3.0.1
reportlab==4.2.5
pypdf==5.1.0
PyMuPDF==1.26.6
scipy==1.11.4
numpy==1.26.2
pyloudnorm==0.1.1
librosa==0.10.1
soundfile==0.12.1
openai==1.57.2
Pillow==10.4.0
ollama

pdf2docx==0.5.8
python-docx==1.1.0

slowapi==0.1.9

# Redis & Celery for async task processing
redis==5.0.1
celery==5.3.6
flower==2.0.1

# Production monitoring & logging
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0
